ChangeLog: Updated ChangeLog. gdip.h: Changed function signature for SeekDelegate function. tiffcodec.c: Minor modification in passing flags to libtiff.
svn path=/trunk/libgdiplus/; revision=30278

+++ src/gdip.h
@@ -1195,7 +1195,7 @@ const EncoderParameter *gdip_find_encoder_parameter (GDIPCONST EncoderParameters
 
 /* Stream handling bits */
 typedef int (*GetBytesDelegate) (unsigned char *, int, BOOL);
-typedef int (*SeekDelegate) (int, int);
+typedef long (*SeekDelegate) (int, int);
 typedef int (*PutBytesDelegate) (unsigned char *, int);
 typedef void (*CloseDelegate) ();
 typedef long (*SizeDelegate) ();
